requirements.txt
Flask
mysql-connector-python
PyYAML